Factor out conversion
