+++ crates/ctt/src/pipeline.rs
@@ -1,4 +1,4 @@
-use crate::conversion_graph::{ConversionGraph, FormatState, build_default_graph, check_lossless};
+use crate::conversion::{ConversionGraph, FormatState, build_default_graph, check_lossless};
 use crate::error::{Error, Result};
 use crate::surface::Image;
 use crate::transforms::Transform;
